requirements.txt
streamlit
phidata
google-generativeai
duckduckgo-search
python-dotenv